src/requirements.txt
google-cloud-language==1.0.2
six